backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
psycopg2-binary>=2.9.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
alembic>=1.11.0
